requirements.txt
discord.py>=2.3.0
openai>=1.0.0
python-dotenv>=1.0.0
sqlalchemy>=2.0.0
alembic>=1.12.0
asyncpg>=0.28.0
aiofiles>=23.0.0
pydantic>=2.0.0
python-dateutil>=2.8.0